backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9        
pydantic==2.8.2
pydantic-settings==2.4.0       
sqlalchemy==2.0.35             
alembic==1.13.2                
asyncpg                
psycopg2-binary       
supabase==2.7.4                
redis==5.0.8
celery[redis]==5.4.0      
flower==2.0.1                  
python-jose[cryptography]==3.3.0   
passlib[bcrypt]==1.7.4             
python-dotenv==1.0.1               
openai==1.46.0                
tiktoken==0.7.0               
langchain==0.3.1               
langchain-openai==0.2.1
langchain-community==0.3.1
websockets==13.0
pypdf==4.3.0                   
python-docx==1.1.2             
pillow==10.4.0                 
boto3==1.35.19                
httpx==0.27.2
beautifulsoup4==4.12.3 
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
apscheduler==3.10.4 
firebase-admin==6.5.0 
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2   
slowapi
redis
supabase
openai
python-dotenv
requests